examples/requirements.txt
absl_py
dm_env
dm_meltingpot
dm_tree
dmlab2d
gymnasium
matplotlib
ml_collections
pettingzoo>=1.22.3
ray[rllib,default]==2.5.0
stable_baselines3
supersuit>=3.7.2
torch
